// playgrounds/react-sandbox/src/core/enums/network/ExplorersEnum.ts
export enum TvmExplorersEnum {
  MAINNET = 'https://tronscan.org/',
  SHASTA = 'https://shasta.tronscan.org/',
}

export enum EvmExplorersEnum {
  BSC_MAINNET = 'https://bscscan.com/',
  BSC_TESTNET = 'https://testnet.bscscan.com/',
  POLYGON_MAINNET = 'https://polygonscan.com/',
  POLYGON_TESTNET = 'https://mumbai.polygonscan.com/',
  AVALANCHE_MAINNET = 'https://snowtrace.io/',
  AVALANCHE_TESTNET = 'https://testnet.snowtrace.io/',
  AURORA_MAINNET = 'https://explorer.mainnet.aurora.dev/',
  AURORA_TESTNET = 'https://explorer.testnet.aurora.dev/',
  ETHEREUM_MAINNET = 'https://etherscan.io/',
  ETHEREUM_GOERLI_TESTNET = 'https://goerli.etherscan.io/',
  MOONBEAM_MAINNET = 'https://moonscan.io/',
  MOONBEAM_TESTNET = 'https://moonbase.moonscan.io/',
  OPTIMISM_MAINNET = 'https://optimistic.etherscan.io/',
  OPTIMISM_TESTNET = 'https://kovan-optimistic.etherscan.io/',
  ARBITRUM_MAINNET = 'https://arbiscan.io/',
  ARBITRUM_TESTNET = 'https://rinkeby-explorer.arbitrum.io/',
}