// playgrounds/react-sandbox/src/core/enums/network/RpcUrlsEnum.ts
export enum TvmGrpsUrlsEnum {
  MAINNET = 'https://api.trongrid.io/',
  SHASTA = 'https://api.shasta.trongrid.io/',
}

export enum EvmRpsUrlsEnum {
  BSC_MAINNET = 'https://bsc-dataseed.binance.org/',
  BSC_TESTNET = 'https://data-seed-prebsc-1-s1.binance.org:8545/',
  POLYGON_MAINNET = 'https://polygon-rpc.com/',
  POLYGON_TESTNET = 'https://matic-mumbai.chainstacklabs.com/',
  AVALANCHE_MAINNET = 'https://api.avax.network/ext/bc/C/rpc/',
  AVALANCHE_TESTNET = 'https://api.avax-test.network/ext/bc/C/rpc',
  AURORA_MAINNET = 'https://mainnet.aurora.dev/',
  AURORA_TESTNET = 'https://testnet.aurora.dev/',
  ETHEREUM_MAINNET = 'https://mainnet.infura.io/v3/',
  ETHEREUM_GOERLI_TESTNET = 'https://rpc.ankr.com/eth_goerli/',
  MOONBEAM_MAINNET = 'https://rpc.api.moonbeam.network/',
  MOONBEAM_TESTNET = 'https://rpc.api.moonbase.moonbeam.network/',
  OPTIMISM_MAINNET = 'https://mainnet.optimism.io/',
  OPTIMISM_TESTNET = 'https://kovan.optimism.io/',
  ARBITRUM_MAINNET = 'https://arb1.arbitrum.io/rpc/',
  ARBITRUM_TESTNET = 'https://rinkeby.arbitrum.io/rpc/',
}